Added New Feature Privating Deleting Message, Selectable 🥳

+++ bot.js
@@ -437,8 +437,8 @@ ${chalk.blue.italic('Made By TOXIC-DEVIL')}`);
                         } else {
                             whats = new Message(conn, msg);
                         }
-
-                        if (command.deleteCommand && msg.key.fromMe) {
+                     
+                        if (config.PVTDELMSG == 'true' && command.deleteCommand && msg.key.fromMe) {
                             await whats.delete(); 
                         }
